# api/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: A RESTful Todo API following Go 2026 Industry Standards.
  version: 2026-04-06

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /todos:
    get:
      summary: List all todos
      operationId: listTodos
      parameters:
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: pageToken
          in: query
          schema:
            type: string
      responses:
        '200':
          description: A list of todos
          headers:
            Total-Count:
              schema:
                type: integer
            Link:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
    post:
      summary: Create a new todo
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'

  /todos/{todoID}:
    parameters:
      - name: todoID
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get a todo by ID
      operationId: getTodo
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
    patch:
      summary: Update a todo
      operationId: updateTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
    delete:
      summary: Delete a todo
      operationId: deleteTodo
      responses:
        '204':
          description: No Content

  /todos/{todoID}:complete:
    post:
      summary: Complete a todo
      operationId: completeTodo
      parameters:
        - name: todoID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'

components:
  schemas:
    Todo:
      type: object
      required:
        - id
        - title
        - completed
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateTodoRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type: string

    UpdateTodoRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string

    Error:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        traceId:
          type: string